feat: change project structure. Cleaner user api. Fixes
Made the project look more like a standalone library instead of a
compleate mess. The whole api is exposed in 2 functions and a is
availabel in 3 lines of code in the main file. Change the inner workings
of the library and how the data is handled in the backgound a lot still
has to be done.