refactor: refactored the underlying class system
Now every kind of handler either class-based or decorator-based is going
to be an ActionHandler instance in the end. This is much more flexible
then the workaround that was present previously. Also, now the Handlers
class is the "primary" class and contains all the logic needed to start
a basic appliation is contained there instead of being scattered all
over the library in different util functions.