feat: basic session broadcasting implementation
Created a raw session broadcasting implementation which allows to send a
message to all websockets connections of the same user which are
currently connected to this server. In the future the API is going to be
refinied as well as adding support for posgres LISTEN/NOTIFY or redis
pub/sub.